Add the new GL_KHR_shader_subgroup extension.
This commit adds the new GL_KHR_shader_subgroup extension, which adds a rich collection of subgroup functionality to GLSL. A subgroup is a set of invocations that can synchronize and share data with each other efficiently.

This extension has corresponding new opcodes in SPIR-V 1.3, and is usable in Vulkan 1.1.