Clarify nature of added fragments to quads
When a primitive only partially covers a quad of fragments, additional
fragment shader invocations are generated to fill it out. According to
other documentation, these are helper invocations. This clarifies that
they are helpers, which suggests how the case where the quad is not
completely composed of covered fragments can be detected.

Resolves issue #77