Change uses of "plane" in the context of clip/cull
Because these distances can be programmed per-vertex in shaders they are
no longer actually planes and phrases like "inside the plane" are
confusing. Replace "plane" with "half-space" because it is clearer what
"inside" and "outside" mean in that context. (Although actually
half-space is not techincally any more accurate than plane here).

See #51.