Update GL_KHR_vulkan_glsl for Rect textures
These are not used in Vulkan because the API does not distinguish
these images as different dimensions, allowing mixing-and-matching with
samplers that do and do not use unnormalized coordinates.

This update removes the additional Rect handles that were added but
could never be used (texture2DRect, etc), explicitly reserves the tokens
that already exist in GLSL (sampler2DRect, etc), and adds a brief
explanation of how to get this same functionality in Vulkan.

Fixes Vulkan-Docs issue #1582.