Split spec into file-per-chapter for Antora
Update spec copyright to current version as found in the Sample-Docs
repository.

Minor markup cleanup.

Aside from the copyright refactoring, this is unchanged relative to
master branch.

Because the copyright appears in its own "Prelude" chapter now, which is
required for inclusion in Antora, the subsequent chapter numbers are all
increased by 1.