Extend grammar tokens with their GLSL representation.
Currently the grammar section of the spec lists all of the tokens
returned by the lexer. It can be difficult to match those lexical tokens
back to the relevant GLSL strings. This CL changes the token list to be
a table, providing the lexical token along with the GLSL which the token
represents.

This makes it easier to map things like `HIGH_PRECISION` to `highp` when
reading the grammar.

The float, integer and boolean lexical tokens are linked back to the
relevant portion of the spec which defines those values.

The `TYPE_NAME` and `FIELD_SELECTION` entries are mapped to `identifier`
as that seems to be what they refer too.