Make `switch_statement_list` and `conditionopt` consistent.
Both `switch_statement_list` and `conditionopt` permit an empty entry.
In one case we used `_nothing_` and the other `_empty`. As well, the
ordering of when the entry was reversed, first in one, second in the
other.

This CL makes them consistent by using `_empty_` and listing it first in
both grammar productions.