Update syntax highlighting
This Cl switches to the `rouge` highlighter which matches the Vulkan
spec. The `rouge-extended-css.rb` file is copied in from the Vulkan-Docs
repo.

The code examples are updated to use `source,glsl` instead of
`source,c++`. This means when mousing over the code blocks in the HTML
version the syntax will correctly be identified as GLSL.